RE-arrange and rename old files (they will be removed later)
